edicion de documento y pdf con link

+++ src/main/java/CONTROLADOR/servlets/sv_documentos.java
@@ -10,6 +10,7 @@ import MODELO.DocumentoDAO;
 import MODELO.FormDoc;
 import MODELO.IdiomaClass;
 import MODELO.TipoClass;
+import MODELO.usuarios.Validador;
 import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -73,10 +74,16 @@ public class sv_documentos extends HttpServlet {
         List<AsignaturaClass> asignaturas = formDoc.obtenerAsignaturas();
         List<IdiomaClass> idiomas = formDoc.obtenerIdiomas();
         List<TipoClass> tipos = formDoc.obtenerTipos();
-        
-        if (asignatura != null) documento.setAsignaturaId(Integer.parseInt(asignatura));
-        if (idioma != null) documento.setIdiomaId(Integer.parseInt(idioma));
-        if (tipo != null) documento.setTipoId(Integer.parseInt(tipo));
+
+        if (asignatura != null) {
+            documento.setAsignaturaId(Integer.parseInt(asignatura));
+        }
+        if (idioma != null) {
+            documento.setIdiomaId(Integer.parseInt(idioma));
+        }
+        if (tipo != null) {
+            documento.setTipoId(Integer.parseInt(tipo));
+        }
 
         List<DocumentoClass> documentos;
         if (asignatura == null && idioma == null && tipo == null) {
@@ -103,29 +110,61 @@ public class sv_documentos extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
+        request.setCharacterEncoding("UTF-8");
+        response.setCharacterEncoding("UTF-8");
         HttpSession session = request.getSession(false);
 
         // VARIABLES
         int idDocumento = Integer.parseInt(request.getParameter("docId"));
         String rutaMiniatura = request.getParameter("rutaMiniatura");
-        
+        String rutaPDF = request.getParameter("rutaPDF");
+        String titulo = request.getParameter("titulo");
+        String autor = request.getParameter("autor");
+        String descripcion = request.getParameter("descripcion");
+        String year = request.getParameter("year");
+
         DocumentoDAO documentoDAO = new DocumentoDAO();
         DocumentoClass documento = new DocumentoClass();
-        documento.setId(idDocumento);
+     
         
-        if (request.getParameter("eliminarDocumento") != null){
+        // VALIDACIONES
+        String errorMessage = Validador.validarTitulo(titulo);
+        if (errorMessage == null) errorMessage = Validador.validarAutor(autor);
+        if (errorMessage == null) errorMessage = Validador.validarDescripcion(descripcion);       
+        if (errorMessage == null) errorMessage = Validador.validarYear(year);        
+
+        if (errorMessage != null) {
+            session.setAttribute("error", errorMessage);
+            response.sendRedirect(request.getHeader("Referer"));
+            return;
+        }
+        
+        documento.setId(idDocumento);
+        documento.setTitulo(titulo);
+        documento.setAutor(autor);
+        documento.setDescripcion(descripcion);
+        documento.setYear(year);
+
+        if (request.getParameter("eliminarDocumento") != null) {
             try {
                 documentoDAO.eliminarDocumento(documento);
-                
+
                 // Eliminar la miniatura del sistema de archivos
                 String rutaContexto = getServletContext().getRealPath("/"); // PROYECTO_BVINEM_JSP\target\PROYECTO_BVINEM-1.0-SNAPSHOT\
                 String rutaAlterada = rutaContexto.replace("\\target\\PROYECTO_BVINEM-1.0-SNAPSHOT\\", "\\src\\main\\webapp");
-                String rutaArchivo = rutaAlterada + File.separator + rutaMiniatura;
+                String rutaArchivoMiniatura = rutaAlterada + File.separator + rutaMiniatura;
+                String rutaArchivoPDF = rutaAlterada + File.separator + rutaPDF;
 
-                File archivoMiniatura = new File(rutaArchivo);
+                File archivoMiniatura = new File(rutaArchivoMiniatura);
                 if (archivoMiniatura.exists()) {
                     archivoMiniatura.delete();
                 }
+
+                File archivoPDF = new File(rutaArchivoPDF);
+                if (archivoPDF.exists()) {
+                    archivoPDF.delete();
+                }
+
                 System.out.println(rutaContexto);
                 session.setAttribute("success", "Documento o libro eliminado exitosamente");
                 response.sendRedirect("sv_documentos");
@@ -133,6 +172,18 @@ public class sv_documentos extends HttpServlet {
                 e.printStackTrace();
                 response.getWriter().print("Error: " + e.getMessage());
                 return;
+            }
+        }
+
+        if (request.getParameter("editDocumento") != null) {
+            try {
+                documentoDAO.editarDocumento(documento);
+                session.setAttribute("success", "Documento actualizado exitosamente");
+                response.sendRedirect("sv_documentos");
+            } catch (SQLException e) {
+                e.printStackTrace();
+                session.setAttribute("error", "Error al actualizar el documento: " + e.getMessage());
+                response.sendRedirect("sv_documentos");
             }
         }
     }
